# Do not allocate so many arrays in optimizer hot path