@aviatesk 2022-03-07
@nanosoldier `runbenchmarks("inference", vs=":master")`