@nanosoldier 2022-03-07
[Your benchmark job](https://github.com/JuliaLang/julia/pull/44492#issuecomment-1060242536) has completed - no performance regressions were detected. A full report can be found [here](https://github.com/JuliaCI/NanosoldierReports/blob/master/benchmark/by_hash/70053ba_vs_811e534/report.md).